Removing `Pride.Core.boolNodeFactory` as logic has been moved to `Pride.FieldTree.FieldBoolean`.
