--- src/Pride/Core/boolNodeFactory.js
@@ -1,28 +1,0 @@
-import nodeFactory from './nodeFactory';
-
-const boolNodeFactory = (type, childTypes) => {
-  return nodeFactory(
-    type,
-    childTypes,
-    function () {
-      if (!['AND', 'OR', 'NOT'].includes(this.value)) {
-        throw new Error('Not a valid boolean value');
-      }
-
-      this.serializedChildren = () => {
-        return this.children.map((child) => {
-          if (child.type === this.type || (child.type === 'literal' && child.value.match(/\s/))) {
-            return `(${child.serialize()})`;
-          }
-          return child.serialize();
-        });
-      };
-
-      this.serialize = () => {
-        return this.serializedChildren().join(` ${this.value} `);
-      };
-    }
-  );
-};
-
-export default boolNodeFactory;
